Implemented 'Update Class Schedules'
Implemented 'Update Class Schedules' function in Admin Functions.

Admins can now View the class schedule, with the most recent at the top of the list.

They can also update the class schedule, selecting the ID of the class, and changing the date+time.

Admins can also add a new class, assigning the class to a trainer. If the trainer is unavailable at the time or has prior commitments, the system will not let them be assigned to the new class.